scripts/tcm-ingest/requirements.txt
httpx>=0.27
pandas>=2.2
pubchempy>=1.0.4
python-dotenv>=1.0
tqdm>=4.66
